refactor(level.py): clarify docstring for level command to specify default behavior when no member is provided
feat(levels.py): consolidate level management commands into a single cog for better organization and maintainability

refactor(levels.py): remove redundant files and integrate functionality into levels cog

fix(levels.py): correct string formatting in logger debug statement

style(help.py): remove unnecessary punctuation in help command aliases field for consistency
